requirements.txt
aiogram>=3.0.0
python-dotenv>=0.20.0
SQLAlchemy>=2.0.0